CESM updates: Remove MPI_BCast from UCX NOXCOEFF_INIT to handle it in CESM code.
This is because NOXCOEFF_INIT is called for all chunks, but not all CPUs in the same communicator have the same amount of chunks when running on the spectral-element dycore. This will result in MPI_BCast hanging as not all CPUs in the communicator are participating in a collective call.
A corresponding update inside CESM geoschem/chemistry.F90 is necessary to restore the original functionality and will be in a companion PR.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>