Fix PBL level calculations
Use local scale height and level thickness to determine
PBL top level and pressure thickness.
The code previously used a fixed global value for scale height,
which is too large near poles and too small in tropics.
This update lowers the PBL top level in the tropics and
raises it near the poles.
The change is typically 1 model level or less.

Signed-off-by: Christopher Holmes <cdholmes@fsu.edu>