Add fixes for DustDead extension settings following TOMAS updates
In setCommonRunSettings.sh for GCHP the code for setting ${RUNDIR_DUSTDEAD_TF}
(i.e. the mass tuning factor for the DustDead extension) was repeated twice:
once for DustDead and once for TOMAS_DustDead. However, the same variables
were used resulting in an error when checking for the DustDead extension
further down in the script. To properly set the mass tuning factor, we now
use a different set of variables in SetCommonRunSettings.sh when when determining
if the DustDead and TOMAS_DustDead extensions are on.

Also, in GCHP's createRunDir.sh TOMAS_SEASALT and TOMAS_DUSTDEAD are
now set to off for benchmark simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>