Move all aerosol_mod gridded module arrays to new derived type in State_Chm
Updates include:
 - New file aermass_container_mod.F90 to hold gridded arrays previously
   in aerosol_mod.F90, e.g. OCFPOA, BCPO, and other aerosol mass arrays
   computed in Aerosol_Conc
 - New derived type AerMassContainer defined in the new file
 - State_Chm%AerMass to point to the AerMassContainer object
 - Code updates to access the aerosol mass arrays via the State_Chm
   pointer, e.g. State_Chm%AerMass%OCFPOA
 - Removal of aerosol_mod.F90 subroutine Cleanup_Aerosol
 - Moved diagnostic submodule Set_AerMass_Diagnostic in aerosol_mod.F90
   to be stored instead in diagnostics_mod.F90
 - Remove the Transport library from GeosCore/CMakeLists.txt

A note on the last update, to remove the Transport library from GeosCore:
This code was put in years ago with the intention of making it easier to
include GEOS-Chem in external models without GEOS-Chem Classic transport.
However, the files in the library are built before GeosCore and include
diagnostics_mod.F90. Excluding this built library in models would prevent
using GEOS-Chem diagnostics. It also prevents using non-transport GeosCore
modules within diagnostics_mod.F90, a problem that came up during this
aerosols update. In practice we do not need this separately built library
when porting GEOS-Chem to other models. There are simpler ways of not
building the transport files for all cases thus far.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>