Remove definition of METDIR from HEMCO_Config.rc for each simulation type
METDIR is not used in the primary HEMCO_Config.rc file for all simulations.
It is however used in the HEMCO_Config.rc.gmao_metfields and
HEMCO_Config.rc.gcap2_metfields files where METDIR is defined again. This
could lead to errors if there were inconsistencies between the two
METDIR definitions. For instance, if users modified the resolution, domain,
or met field type in the HEMCO_Config.rc.*metfields file and not in the
primary HEMCO_Config.rc file then HEMCO would crash because of it would use
the incorrect METDIR path. This is resolved by removing the redundant
definition in the primary HEMCO_Config.rc files and just defining it in
HEMCO_Config.rc.*metfields where it is actually used.

This update also allows us to remove METDIR from the GCHP HEMCO_Config.rc
files where it is not used because met fields defined in ExtData.rc and
read in by MAPL, not HEMCO.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>