Modify HEMCO_Config.rc.gmao_metfields_0125 to use 0.25x0.3125 NA files for non-dynamic fields
The 0.125x0.15625 files are only available for dynamic collections A1dyn
and I1dyn. The remaining met fields when running at this resolution should
be read at 0.25x0.3125. This file is currently hardcoded to use the North
America (NA) files at 0.25x0.3125 resolution for testing, but this should
be made more flexible in the future.

The 0.125x0.15625 files have also been updated here so that daily files are
used instead of hourly files. The use of daily 24-hour met field files has
already been implemented in GEOS-Chem, but it does not currently know how
to handle hourly files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>