Restrict Cloud-J aerosol conc assignment to fullchem/aerosol simulations
Aerosol values are only available when using the fullchem or aerosol
simulations. The mercury simulation uses Cloud-J photolysis but
aerosol concentrations are not available to use within it. Values are
thus kept at zero for the mercury simulation. This is consistent with
legacy Fast-JX in GEOS-Chem which took optical depth rather than
concentration as input. Aerosol optical depths, including soil dust,
were only compute for fullchem and aerosol simulations and left as zero
for mercury.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>