Fix uptake coefficients on stratospheric aerosol
Corrected the formula for 1st order
heterogeneous chemical loss on stratospheric aerosol for a few species.
The affected species are NO2, NO3, and VOC that experience heterogeneous loss on aerosol.
The fix should increase heterogeneous losses of these species in the stratosphere.
The overall impact on VOC should be small as there is little VOC in the stratosphere.
The additional loss of NO2 and NO3 should also increase stratospheric HNO3.