Update integration tests to pass quick option to compile scripts
The quick option included in the integration tests generates and runs a
subset of simulations. However, with this option activated, executables
were still being created for all possible compile options. We now pass
the quick argument to integrationTestCompile.sh so that only the necessary
executables are built. This will further speed up the quick integration tests.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>