Retire tagCH4 simulation as an option in GCClassic
IMPORTANT: This update impacts simulation numbers in createRunDir.sh for
GCClassic.

The tagged CH4 simulation has been removed as an option for GCClassic. This
simulation is sparsely used. Also, the CH4, CO2, and tagCO simulations are
slated for removal in an upcoming version in favor of the carbon simulation
which can also be run in single tracer mode. If users want to utilize tagged
CH4 tracers in the carbon simulation, they will need to make the modifications
to the input files and chemistry mechanism.

As noted above this update also reorders the simulation numbers when executing
./createRunDir for GCClassic. The carbon simulation has been bumped up the list
to number 3 to confirm with alphabetical order. The CH4, CO2, and tagCO
options have been moved to the end of the list in preparation for their
eventual removal. The scripts that utilize simulation number (e.g.
integrationTestCreate.sh and parallelTestCreate.sh for GCClassic) have also
been updated. Any external user scripts that hardcode simulation number will
need to be updated. For example, this will impact the IMI.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>