Add fix to singleCarbonSpecies.sh to avoid removing CEDS entry for CO2-only simulations
The carbon simulation with CO2 only was failing integration tests due to a
failed ExtData import for CEDS_CO2_SHIP. The sed statements in singleCarbonSpecies.sh
removed any entries with "CEDS_CO" if the CO species was not included in
the simulation, also inadvertently removing the CEDS_CO2_SHIP entry. This
has now been fixed by adding an underscore after CEDS_CO2. Other lines in
that sed statement have also been updated to include an underscore or trailing
white space to avoid this issue in the future.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>