Fix typos in HEMCO_Config.rc.CH4 causing double counting of mobile combustion emissions
The emission for GHGI_EE_COAST_OTHER_MCOMB in HEMCO_Config.rc.CH4 was incorrectly
assigned a hierarcy of 100 (added to regional US emissions) when it should
have been 1 (added to EDGAR global emissions). The wrong mask was also applied
to this field - it should be 1009 (CONUS_MASK_MIRROR) and not 1008 (CONUS_MASK).
These typos caused mobile combustion emissions to be double counted over the US.
The typos were not present in the HEMCO_Config.rc files for carbon simulations.

Also made other white space and comment updates to HEMCO_Config.rc files
for the carbon and CH4 simulations to ensure consistency between the two
simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>