Modify tagCO simulation to use GFED emissions and v5 OH for validation against carbon simulation
The carbon simulation currently uses GFED biomass burning emissions by default
(for all carbon species). It also uses GEOS-Chem v5 OH which is currently
required for CH4 in the carbon simulation. To simplify validation of
the CO in the tagCO simulation vs the carbon simulation, we update the
HEMCO_Config.rc.tagCO to use the same options.

Also made a minor change to the OCS biomass burning diagnostic name for
consistency with the other carbon species.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>