Remove setting DELP_DRY to zero when not found in restart file
The code for initializing DELP_DRY was moved from hco_interface_gc_mod.F90
to hco_utilities_gc_mod.F90 in PR #2521. This inadvertently led to negative
concentrations in simulations where DELP_DRY was not found in the restart file.
This is because DELP_DRY is updated in the first call to AirQnt in
hco_interfaces_gc_mod.F90. When it is not found in the restart file, it is
reset to zero which causes issues later in the simulation (manifesting as
negative values in mixing). To avoid these issues, we remove the ELSE
statement setting DELP_DRY to zero when it's not found in the restart file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>