Add option to read GC-Classic restart locally as real8
This option is only available for simulations other than nested and Hg.
It is a hidden feature, meaning the entry for it is not included in
geoschem_config.yml file templates. Users may add simulation menu entry
read_restart_as_real8 to access the feature. Setting that to true will
read the GC-Classic restart file locally in GEOS-Chem as REAL8 and use
those values rather than the values read by HEMCO which are in REAL4.

This update is to facilitate testing bit-for-bit reproducibility of
GC-Classic when breaking up a run in time. Ideally HEMCO would be able
to read file data as REAL8 but that feature has not yet been implemented.
Once it is the capability to read the GC-Classic restart file locally
in GEOS-Chem can be removed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>