Pass in second when calling Accept_External_Date_Time
This is a fix to the GEOS-Chem/GCHP chunk interface to pass in the
"second" information to Accept_External_Date_Time from GCHP_Chunk_Run.

The omission of this argument caused a bug within WRF-GC, which uses
a fork of this interface, to have the incorrect time stamp when calling
HEMCO after updating the GEOS-Chem clock on fractional minutes (i.e.,
when seconds /= 0). This would cause the HEMCO clock to "rewind" itself
from the correct time to hh:mm:00 and lose the "first timestep" flag,
necessary for loading information on MEGAN biogenic emissions and possibly
other initialization subroutines in HEMCO.

This update should not impact other models using GEOS-Chem, but is being
contributed here in order to keep the interface robust.