Update GCHP mass flux and courant number import scaling calculation
If using mass flux and courant number imports from the meteorology archive
then then these imports must be scaled using factors dependent on run
timestep, run grid resolution, timestep used to produce the met archive,
and grid resolution used to produce the met archive. These scale factors
were introduced in 14.6.3 but must be changed again in 14.7.0 given
changes in MAPL horizontal flux regridding in that version. The new
scaling is the same as the old but divided by an additional quantity of
(met grid resolution / run grid resolution). This results in cancelling
out the grid-dependency for mass flux and squaring the grid dependency
of courant numbers.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>