Fix typos causing carbon simulations with the Carbon collection enabled to crash
When enabling the Carbon collection in HISTORY.rc, GEOS-Chem crashes with
an out-of-bounds error in KPP/carbon/carbon_Funcs.F90. This can be tracked
to the uses of incorrect ind_ variable names -- ind_COfromCH4 should be
ind_PCOfromCH4 and ind_COfromNMVOC should be ind_PCOfromNMVOC to reflect
the species name change introduced in pull request #2800.

We now also turn on the Carbon collection in HISTORY.rc by default to
catch potential bugs like this during integration tests.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>